+++ ta4j-core/src/main/java/org/ta4j/core/BaseBarSeriesBuilder.java
@@ -29,6 +29,7 @@ import java.util.List;
 import org.ta4j.core.bars.TimeBarBuilderFactory;
 import org.ta4j.core.num.DecimalNumFactory;
 import org.ta4j.core.num.NumFactory;
+import org.ta4j.core.utils.BarUtil;
 
 /**
  * A builder to build a new {@link BaseBarSeries}.
@@ -68,11 +69,7 @@ public class BaseBarSeriesBuilder implements BarSeriesBuilder {
 
             if (!isNumFactoryAssigned) {
                 // use numFactory derived from bars instead of default numFactory
-                var closePrice = bars.get(0).getClosePrice();
-                if (closePrice != null) {
-                    var derivedNumFactory = closePrice.getNumFactory();
-                    numFactory = derivedNumFactory;
-                }
+                numFactory = BarUtil.selectNumFactoryForBar(bars.getFirst());
             }
 
             // check if each bar has the same numFactory as the series numFactory